requirements.txt
aiohttp>=3.9.0
aiodns>=3.1.0
dnspython>=2.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
rich>=13.7.0
tqdm>=4.66.0
